Fix all clippy warnings across workspace
Resolve 66 clippy warnings including: unnecessary casts, redundant
closures, manual Range::contains, redundant pattern matching,
io_other_error, manual_div_ceil, manual_is_multiple_of, ptr_arg
(&PathBuf -> &Path), field_reassign_with_default, and others. Add
#[allow] for intentional patterns (too_many_arguments, result_unit_err,
declare_interior_mutable_const). Add Default impl for VmxContext.

+++ crates/bedrock-vmx/src/exits/misc.rs
@@ -293,7 +293,7 @@ pub fn dump_triple_fault_state<C: VmContext>(ctx: &C) {
 pub fn handle_xsetbv<C: VmContext>(ctx: &mut C) -> ExitHandlerResult {
     let gprs = ctx.state().gprs;
     let xcr_num = gprs.rcx as u32;
-    let value = ((gprs.rdx as u64) << 32) | (gprs.rax as u64 & 0xFFFFFFFF);
+    let value = (gprs.rdx << 32) | (gprs.rax & 0xFFFFFFFF);
 
     // Only XCR0 is currently defined
     if xcr_num != 0 {
